+++ package.json
@@ -15,6 +15,7 @@
     "react": "^17.0.2",
     "react-bootstrap": "^2.1.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.2",
     "react-scripts": "4.0.3",
     "typescript": "^4.6.2",
     "web-vitals": "^2.1.4"
