+++ src/screens/SendTransaction.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function SendTransaction() {
+
+  return (
+    <div className="container">
+      <h1>SendTransaction</h1>
+    </div>
+  );
+}
+
+export default SendTransaction;
